Revert "Revert "Hierarchy organised""
