+++ client/main.js
@@ -1,6 +1,7 @@
 import { Template } from 'meteor/templating';
 import { ReactiveVar } from 'meteor/reactive-var';
 
-import '../imports/ui/feed.js';
-import '../imports/ui/attendence.js';
-import '../imports/ui/body.js';
+import '../imports/ui/Others/feed.js';
+import '../imports/ui/Facilitator/attendence.js';
+import '../imports/ui/Others/body.js';
+// import '../imports/ui/Others/routes.js';
