+++ imports/ui/Facilitator/attendence.html
@@ -6,14 +6,14 @@
 </head> -->
 {{> def}}
   {{#basicTabs name="temp" tabs=tabs}}
-
-  {{#tabContent slug="session1"}}
+  
+  {{#tabContent slug="session1"}} 
   <body>
 
       <header>
          IPAN
          <label class="hide-completed">
-              <input type="checkbox"/>
+              <input type="checkbox"/> 
               Show Absent
       </label>
           <label class="show-completed">
@@ -60,7 +60,7 @@
           </label>
           <ul>
               {{#each tasks1}}
-                  {{> task1}}
+                  {{> task1}} 
               {{/each}}
               <!-- {{> yield}} -->
           </ul>
