+++ imports/ui/Others/feed.html
@@ -18,17 +18,19 @@
   				<span class="icon-bar"></span>
   				<span class="icon-bar"></span>
   			</button>
-  			<a class="navbar-brand" href="#">User 1</a>
+  			<a class="navbar-brand" href="#"><span style="font-size:16px;" class="pull-center hidden-xs showopacity glyphicon glyphicon-user"></span></a>
   		</div>
   		<!-- Collect the nav links, forms, and other content for toggling -->
   		<div class="collapse navbar-collapse" id="bs-sidebar-navbar-collapse-1">
   			<ul class="nav navbar-nav">
           <li class="dropdown">
-  					<a href="#" class="dropdown-toggle" data-toggle="dropdown">Sessions<span style="font-size:16px;" class="pull-right hidden-xs-down showopacity glyphicons glyphicons-address-book"></span></a>
+  					<a href="#" class="dropdown-toggle" data-toggle="dropdown">Sessions<span style="font-size:16px;" class="pull-right hidden-xs-down showopacity glyphicon glyphicon-tasks"></span></a>
   					<ul class="dropdown-menu forAnimate" role="menu">
-  						<li><a href="{{pathFor route='attendence'}}">Sunshine</a></li>
-  						<li><a href="#">Burwood</a></li>
-  						<li><a href="#">Docklands</a></li>
+						  <li><a href="{{pathFor route='attendence'}}">Sunshine<span class="pull-right hidden-xs showopacity glyphicon glyphicon-map-marker"></span></a></li>
+						 
+						  <li><a href="#">Burwood<span class="pull-right hidden-xs showopacity glyphicon glyphicon-map-marker"></span></a></li>
+						  
+  						<li><a href="#">Docklands<span class="pull-right hidden-xs showopacity glyphicon glyphicon-map-marker"></span></a></li>
   						<!-- <li class="divider"></li>
   						<li><a href="#">Separated link</a></li>
   						<li class="divider"></li>
@@ -37,7 +39,8 @@
   				</li>
   				<!-- <li class="active"><a href="{{pathFor route='attendence'}}" class="dropdown-toggle" data-toggle="dropdown">Home<span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-home carret"></span></a></li> -->
   				<li ><a href="#">Profile<span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-user"></span></a></li>
-  				<li ><a href="#">Messages<span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-envelope"></span></a></li>
+				  <li ><a href="#">Messages<span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-envelope"></span></a></li>
+				  <li><a href="#">Settings<span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-cog"></span></a></li>
   				<!-- <li class="dropdown">
   					<a href="#" class="dropdown-toggle" data-toggle="dropdown">Settings <span class="caret"></span><span style="font-size:16px;" class="pull-right hidden-xs showopacity glyphicon glyphicon-cog"></span></a>
   					<ul class="dropdown-menu forAnimate" role="menu">
