+++ imports/ui/Others/feed.js
@@ -1,7 +1,10 @@
 import './feed.html';
-import '../routes.js';
+import './routes.js';
 import './body.js';
-import './top.html';
+// import '../Facilitator/attendence.js';
+// import '../Facilitator/attendence.html';
+
+
 function htmlbodyHeightUpdate(){
 		var height3 = $( window ).height()
 		var height1 = $('.nav').height()+50
@@ -15,7 +18,7 @@ function htmlbodyHeightUpdate(){
 			$('html').height(Math.max(height1,height3,height2));
 			$('body').height(Math.max(height1,height3,height2));
 		}
-
+		
 	}
 	$(document).ready(function () {
 		htmlbodyHeightUpdate()
