+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@babel/runtime": "7.0.0-beta.55",
+    "boostrap": "^2.0.0",
     "bootstrap": "^4.1.3",
     "meteor-node-stubs": "^0.4.1"
   },
