+++ server/main.js
@@ -1,4 +1,4 @@
-import '../imports/routes.js';
+import '../imports/ui/Others/routes.js';
 //import '../client/main.js';
 
 import { Tasks } from '../imports/api/task.js';
